ddd/pyproject.toml
[project]
name = "ddd"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio (>=1.73.1,<2.0.0)",
    "grpcio-tools (>=1.73.1,<2.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "mypy (>=1.16.1,<2.0.0)",
    "ruff (>=0.12.3,<0.13.0)",
    "mysql-connector-python (>=9.3.0,<10.0.0)"
]

[tool.poetry]
packages = [{include = "ddd", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
